[BE] refactor : 컨트롤러 통합
- ApiController.java : ApiController로 통합했습니다.
- url 관련 상수들을 의미에 맞게 ApiUrl 옮겼습니다.

associate with #17
